:truck: rename HashtagBadge
